API compatible Serve for easy migration
